Renamed console command classes
